requirements.txt
2captcha-python==1.1.0
python-decouple==3.4
requests==2.25.1
